AI/Joochan/BE/requirements_1.txt
APScheduler
boto3==1.38.10
numpy==1.24.4
Cython
cmake==4.0.0
fastapi
gdown
pillow
pycryptodome
qdrant-client
uvicorn
websockets
opencv-python
opencv-contrib-python
python-dotenv
pytz
setuptools
requests
public-ip